Update NuGet Metadata
After speaking with the NuGet team regarding Cake Addin's, there were a
couple of recommendations regarding adding some additional metadata.
This addresses this by adding a license file to the repository and then using a
license expression in the csproj file to reflect this on NuGet.org.  I have taken the liberty
of assuming MIT, if this is not the case, I can correct.  Also, the
RepositoryUrl has been included as well.